Remove external dependency, finalize expandable card
