--- packages/app/lib/ui/widgets/expansion_tile_custom.dart
@@ -1,629 +1,0 @@
-import 'package:flutter/material.dart';
-
-const Duration _kExpand = Duration(milliseconds: 200);
-
-/// A single-line [ListTile] with an expansion arrow icon that expands or collapses
-/// the tile to reveal or hide the [children].
-///
-/// This widget is typically used with [ListView] to create an
-/// "expand / collapse" list entry. When used with scrolling widgets like
-/// [ListView], a unique [PageStorageKey] must be specified to enable the
-/// [ExpansionTileCustom] to save and restore its expanded state when it is scrolled
-/// in and out of view.
-///
-/// This class overrides the [ListTileThemeData.iconColor] and [ListTileThemeData.textColor]
-/// theme properties for its [ListTile]. These colors animate between values when
-/// the tile is expanded and collapsed: between [iconColor], [collapsedIconColor] and
-/// between [textColor] and [collapsedTextColor].
-///
-/// The expansion arrow icon is shown on the right by default in left-to-right languages
-/// (i.e. the trailing edge). This can be changed using [controlAffinity]. This maps
-/// to the [leading] and [trailing] properties of [ExpansionTile].
-///
-/// {@tool dartpad}
-/// This example demonstrates different configurations of ExpansionTile.
-///
-/// ** See code in examples/api/lib/material/expansion_tile/expansion_tile.0.dart **
-/// {@end-tool}
-///
-/// See also:
-///
-///  * [ListTile], useful for creating expansion tile [children] when the
-///    expansion tile represents a sublist.
-///  * The "Expand and collapse" section of
-///    <https://material.io/components/lists#types>
-class ExpansionTileCustom extends StatefulWidget {
-  const ExpansionTileCustom(
-      {Key? key,
-      this.leading,
-      required this.title,
-      this.subtitle,
-      this.onExpansionChanged,
-      this.children = const <Widget>[],
-      this.trailing,
-      this.initiallyExpanded = false,
-      this.maintainState = false,
-      this.tilePadding,
-      this.expandedCrossAxisAlignment,
-      this.expandedAlignment,
-      this.childrenPadding,
-      this.backgroundColor,
-      this.collapsedBackgroundColor,
-      this.textColor,
-      this.collapsedTextColor,
-      this.iconColor,
-      this.collapsedIconColor,
-      this.controlAffinity,
-      this.clipBehavior,
-      this.decoration,
-      this.borderRadius,
-      this.border,
-      this.boxShadow,
-      this.collapsedBorderColor,
-      this.expendedBorderColor,
-      this.isHasTopBorder = true,
-      this.isHasBottomBorder = true,
-      this.isHasLeftBorder = false,
-      this.isHasRightBorder = false,
-      this.isHasTrailing = true,
-      this.isEnableExpanded = true,
-      this.isDefaultVerticalPadding = true})
-      : assert(
-          expandedCrossAxisAlignment != CrossAxisAlignment.baseline,
-          'CrossAxisAlignment.baseline is not supported since the expanded children '
-          'are aligned in a column, not a row. Try to use another constant.',
-        ),
-        assert(isHasTrailing == true || trailing == null,
-            'Not necessary `trailing` when `isHasTrailing` is false'),
-        assert(
-            (boxShadow != null &&
-                    backgroundColor != null &&
-                    collapsedBackgroundColor != null) ||
-                boxShadow == null,
-            'By default the `backgroundColor` and `collapseBackgroundColor` is transparent, so you need add value when define `boxShadow`'),
-        assert(
-            (borderRadius != null &&
-                    ((isHasTopBorder &&
-                            isHasBottomBorder &&
-                            isHasLeftBorder &&
-                            isHasRightBorder) ||
-                        (!isHasTopBorder &&
-                            !isHasBottomBorder &&
-                            !isHasLeftBorder &&
-                            !isHasRightBorder))) ||
-                borderRadius == null,
-            'A borderRadius can only be given for a uniform Border, \n more info here: https://stackoverflow.com/questions/58812778/a-borderradius-can-only-be-given-for-uniform-borders'),
-        super(key: key);
-
-  /// A widget to display before the title.
-  ///
-  /// Typically a [CircleAvatar] widget.
-  ///
-  /// Note that depending on the value of [controlAffinity], the [leading] widget
-  /// may replace the rotating expansion arrow icon.
-  final Widget? leading;
-
-  /// The primary content of the list item.
-  ///
-  /// Typically a [Text] widget.
-  final Widget title;
-
-  /// Additional content displayed below the title.
-  ///
-  /// Typically a [Text] widget.
-  final Widget? subtitle;
-
-  /// Called when the tile expands or collapses.
-  ///
-  /// When the tile starts expanding, this function is called with the value
-  /// true. When the tile starts collapsing, this function is called with
-  /// the value false.
-  final ValueChanged<bool>? onExpansionChanged;
-
-  /// The widgets that are displayed when the tile expands.
-  ///
-  /// Typically [ListTile] widgets.
-  final List<Widget> children;
-
-  /// The color to display behind the sublist when expanded.
-  ///
-  /// If this property is null then [ExpansionTileThemeData.backgroundColor] is used. If that
-  /// is also null then Colors.transparent is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? backgroundColor;
-
-  /// When not null, defines the background color of tile when the sublist is collapsed.
-  ///
-  /// If this property is null then [ExpansionTileThemeData.collapsedBackgroundColor] is used.
-  /// If that is also null then Colors.transparent is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? collapsedBackgroundColor;
-
-  /// A widget to display after the title.
-  ///
-  /// Note that depending on the value of [controlAffinity], the [trailing] widget
-  /// may replace the rotating expansion arrow icon.
-  final Widget? trailing;
-
-  /// Specifies if the list tile is initially expanded (true) or collapsed (false, the default).
-  final bool initiallyExpanded;
-
-  /// Specifies whether the state of the children is maintained when the tile expands and collapses.
-  ///
-  /// When true, the children are kept in the tree while the tile is collapsed.
-  /// When false (default), the children are removed from the tree when the tile is
-  /// collapsed and recreated upon expansion.
-  final bool maintainState;
-
-  /// Specifies padding for the [ListTile].
-  ///
-  /// Analogous to [ListTile.contentPadding], this property defines the insets for
-  /// the [leading], [title], [subtitle] and [trailing] widgets. It does not inset
-  /// the expanded [children] widgets.
-  ///
-  /// If this property is null then [ExpansionTileThemeData.tilePadding] is used. If that
-  /// is also null then the tile's padding is `EdgeInsets.symmetric(horizontal: 16.0)`.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final EdgeInsetsGeometry? tilePadding;
-
-  /// Specifies the alignment of [children], which are arranged in a column when
-  /// the tile is expanded.
-  ///
-  /// The internals of the expanded tile make use of a [Column] widget for
-  /// [children], and [Align] widget to align the column. The `expandedAlignment`
-  /// parameter is passed directly into the [Align].
-  ///
-  /// Modifying this property controls the alignment of the column within the
-  /// expanded tile, not the alignment of [children] widgets within the column.
-  /// To align each child within [children], see [expandedCrossAxisAlignment].
-  ///
-  /// The width of the column is the width of the widest child widget in [children].
-  ///
-  /// If this property is null then [ExpansionTileThemeData.expandedAlignment]is used. If that
-  /// is also null then the value of `expandedAlignment` is [Alignment.center].
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Alignment? expandedAlignment;
-
-  /// Specifies the alignment of each child within [children] when the tile is expanded.
-  ///
-  /// The internals of the expanded tile make use of a [Column] widget for
-  /// [children], and the `crossAxisAlignment` parameter is passed directly into the [Column].
-  ///
-  /// Modifying this property controls the cross axis alignment of each child
-  /// within its [Column]. Note that the width of the [Column] that houses
-  /// [children] will be the same as the widest child widget in [children]. It is
-  /// not necessarily the width of [Column] is equal to the width of expanded tile.
-  ///
-  /// To align the [Column] along the expanded tile, use the [expandedAlignment] property
-  /// instead.
-  ///
-  /// When the value is null, the value of `expandedCrossAxisAlignment` is [CrossAxisAlignment.center].
-  final CrossAxisAlignment? expandedCrossAxisAlignment;
-
-  /// Specifies padding for [children].
-  ///
-  /// If this property is null then [ExpansionTileThemeData.childrenPadding] is used. If that
-  /// is also null then the value of `childrenPadding` is [EdgeInsets.zero].
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final EdgeInsetsGeometry? childrenPadding;
-
-  /// The icon color of tile's expansion arrow icon when the sublist is expanded.
-  ///
-  /// Used to override to the [ListTileThemeData.iconColor].
-  ///
-  /// If this property is null then [ExpansionTileThemeData.iconColor] is used. If that
-  /// is also null then the value of [ListTileThemeData.iconColor] is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? iconColor;
-
-  /// The icon color of tile's expansion arrow icon when the sublist is collapsed.
-  ///
-  /// Used to override to the [ListTileThemeData.iconColor].
-  final Color? collapsedIconColor;
-
-  /// The color of the tile's titles when the sublist is expanded.
-  ///
-  /// Used to override to the [ListTileThemeData.textColor].
-  ///
-  /// If this property is null then [ExpansionTileThemeData.textColor] is used. If that
-  /// is also null then the value of [ListTileThemeData.textColor] is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? textColor;
-
-  /// The color of the tile's titles when the sublist is collapsed.
-  ///
-  /// Used to override to the [ListTileThemeData.textColor].
-  ///
-  /// If this property is null then [ExpansionTileThemeData.collapsedTextColor] is used. If that
-  /// is also null then the value of [ListTileThemeData.textColor] is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? collapsedTextColor;
-
-  /// Typically used to force the expansion arrow icon to the tile's leading or trailing edge.
-  ///
-  /// By default, the value of `controlAffinity` is [ListTileControlAffinity.platform],
-  /// which means that the expansion arrow icon will appear on the tile's trailing edge.
-  final ListTileControlAffinity? controlAffinity;
-
-  ///The decoration to paint behind the child.
-  /// Use the color property to specify a simple solid color.
-  /// The child is not clipped to the decoration. To clip a child to the shape of a particular ShapeDecoration, consider using a ClipPath widget.
-  final Decoration? decoration;
-
-  /// If non-null, the corners of this box are rounded by this [BorderRadius].
-  ///
-  /// Applies only to boxes with rectangular shapes; ignored if [shape] is not
-  /// [BoxShape.rectangle].
-  ///
-  /// {@macro flutter.painting.BoxDecoration.clip}
-  final BorderRadius? borderRadius;
-
-  /// A border to draw above the background [color], [gradient], or [image].
-  ///
-  /// Follows the [shape] and [borderRadius].
-  ///
-  /// Use [Border] objects to describe borders that do not depend on the reading
-  /// direction.
-  ///
-  /// Use [BoxBorder] objects to describe borders that should flip their left
-  /// and right edges based on whether the text is being read left-to-right or
-  /// right-to-left.
-  final BoxBorder? border;
-
-  ///The clip behavior when Container.decoration is not null.
-  /// Defaults to Clip.none. Must be Clip.none if decoration is null.
-  /// If a clip is to be applied, the Decoration.getClipPath method for the provided decoration must return a clip path.
-  /// (This is not supported by all decorations; the default implementation of that method throws an UnsupportedError.)
-  final Clip? clipBehavior;
-
-  ///A list of shadows cast by this box behind the box.
-  /// The shadow follows the shape of the box.
-  /// See also:
-  /// kElevationToShadow, for some predefined shadows used in Material Design.
-  /// PhysicalModel, a widget for showing shadows.
-  final List<BoxShadow>? boxShadow;
-
-  /// The color to display border box when collapsed.
-  ///
-  /// If this property is null then [ExpansionTileThemeData.backgroundColor] is used. If that
-  /// is also null then Colors.transparent is used.
-  ///
-  /// See also:
-  ///
-  /// * [ExpansionTileTheme.of], which returns the nearest [ExpansionTileTheme]'s
-  ///   [ExpansionTileThemeData].
-  final Color? collapsedBorderColor;
-
-  /// The color to display border box when expanded.
-  ///
-  /// If this property is null then [ThemeData.dividerColor] is used. If that
-  /// is also null then Colors.transparent is used.
-  ///
-  /// See also:
-  ///
-  /// * [ThemeData.of], which returns the nearest [ThemeData].
-  final Color? expendedBorderColor;
-
-  final bool isHasTopBorder;
-  final bool isHasBottomBorder;
-  final bool isHasLeftBorder;
-  final bool isHasRightBorder;
-
-  final bool isHasTrailing;
-
-  ///The widget can expand or NOT.
-  ///It helps when you want to prohibit expansion until the API requests success.
-  ///More detail: https://github.com/congthanhng/Expansion-Tile-Group/issues/22
-  final bool isEnableExpanded;
-
-  ///Remove completely default vertical title padding
-  final bool isDefaultVerticalPadding;
-
-  @override
-  State<ExpansionTileCustom> createState() => ExpansionTileCustomState();
-}
-
-class ExpansionTileCustomState extends State<ExpansionTileCustom>
-    with SingleTickerProviderStateMixin {
-  static final Animatable<double> _easeOutTween =
-      CurveTween(curve: Curves.easeOut);
-  static final Animatable<double> _easeInTween =
-      CurveTween(curve: Curves.easeIn);
-  static final Animatable<double> _halfTween =
-      Tween<double>(begin: 0.0, end: 0.5);
-
-  final ColorTween _borderColorTween = ColorTween();
-  final ColorTween _headerColorTween = ColorTween();
-  final ColorTween _iconColorTween = ColorTween();
-  final ColorTween _backgroundColorTween = ColorTween();
-
-  late AnimationController _controller;
-  late Animation<double> _iconTurns;
-  late Animation<double> _heightFactor;
-  late Animation<Color?> _borderColor;
-  late Animation<Color?> _headerColor;
-  late Animation<Color?> _iconColor;
-  late Animation<Color?> _backgroundColor;
-
-  bool _isExpanded = false;
-
-  @override
-  void initState() {
-    super.initState();
-    _controller = AnimationController(duration: _kExpand, vsync: this);
-    _heightFactor = _controller.drive(_easeInTween);
-    _iconTurns = _controller.drive(_halfTween.chain(_easeInTween));
-    _borderColor = _controller.drive(_borderColorTween.chain(_easeOutTween));
-    _headerColor = _controller.drive(_headerColorTween.chain(_easeInTween));
-    _iconColor = _controller.drive(_iconColorTween.chain(_easeInTween));
-    _backgroundColor =
-        _controller.drive(_backgroundColorTween.chain(_easeOutTween));
-
-    if (_isExpanded) {
-      _controller.value = 1.0;
-    }
-  }
-
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
-  }
-
-  void expand() {
-    _setExpanded(true);
-  }
-
-  void collapse() {
-    _setExpanded(false);
-  }
-
-  void toggle() {
-    _setExpanded(!_isExpanded);
-  }
-
-  void expandedChanged(bool isExpanded) {
-    _setExpanded(isExpanded);
-  }
-
-  void _setExpanded(bool isExpanded) {
-    if (!isExpanded && !widget.isEnableExpanded) return;
-    if (_isExpanded != isExpanded) {
-      setState(() {
-        _isExpanded = isExpanded;
-        if (_isExpanded) {
-          _controller.forward();
-        } else {
-          _controller.reverse().then((_) {
-            setState(() {
-              // Rebuild without widget.children.
-            });
-          });
-        }
-        PageStorage.of(context).writeState(context, _isExpanded);
-      });
-      if (widget.onExpansionChanged != null) {
-        widget.onExpansionChanged!(_isExpanded);
-      }
-    }
-  }
-
-  void _handleTap() {
-    if (!_isExpanded && !widget.isEnableExpanded) return;
-    setState(() {
-      _isExpanded = !_isExpanded;
-      if (_isExpanded) {
-        _controller.forward();
-      } else {
-        _controller.reverse().then<void>((void value) {
-          if (!mounted) {
-            return;
-          }
-          setState(() {
-            // Rebuild without widget.children.
-          });
-        });
-      }
-      PageStorage.of(context).writeState(context, _isExpanded);
-    });
-    widget.onExpansionChanged?.call(_isExpanded);
-  }
-
-  // Platform or null affinity defaults to trailing.
-  ListTileControlAffinity _effectiveAffinity(
-      ListTileControlAffinity? affinity) {
-    switch (affinity ?? ListTileControlAffinity.trailing) {
-      case ListTileControlAffinity.leading:
-        return ListTileControlAffinity.leading;
-      case ListTileControlAffinity.trailing:
-      case ListTileControlAffinity.platform:
-        return ListTileControlAffinity.trailing;
-    }
-  }
-
-  Widget? _buildIcon(BuildContext context) {
-    return RotationTransition(
-      turns: _iconTurns,
-      child: const Icon(Icons.expand_more),
-    );
-  }
-
-  Widget? _buildLeadingIcon(BuildContext context) {
-    if (_effectiveAffinity(widget.controlAffinity) !=
-        ListTileControlAffinity.leading) {
-      return null;
-    }
-    return _buildIcon(context);
-  }
-
-  Widget? _buildTrailingIcon(BuildContext context) {
-    if (_effectiveAffinity(widget.controlAffinity) !=
-        ListTileControlAffinity.trailing) {
-      return null;
-    }
-    return _buildIcon(context);
-  }
-
-  Widget _buildChildren(BuildContext context, Widget? child) {
-    final ExpansionTileThemeData expansionTileTheme =
-        ExpansionTileTheme.of(context);
-    final Color borderSideColor = _borderColor.value ?? Colors.transparent;
-
-    return Container(
-      decoration: widget.decoration ??
-          BoxDecoration(
-            color: _backgroundColor.value ??
-                expansionTileTheme.backgroundColor ??
-                Colors.transparent,
-            border: (widget.border ??
-                Border(
-                    top: widget.isHasTopBorder
-                        ? BorderSide(color: borderSideColor)
-                        : const BorderSide(color: Colors.transparent),
-                    bottom: widget.isHasBottomBorder
-                        ? BorderSide(color: borderSideColor)
-                        : const BorderSide(color: Colors.transparent),
-                    right: widget.isHasRightBorder
-                        ? BorderSide(color: borderSideColor)
-                        : const BorderSide(color: Colors.transparent),
-                    left: widget.isHasLeftBorder
-                        ? BorderSide(color: borderSideColor)
-                        : const BorderSide(color: Colors.transparent))),
-            borderRadius: widget.borderRadius,
-            boxShadow: widget.boxShadow,
-          ),
-      clipBehavior: widget.clipBehavior ?? Clip.hardEdge,
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          ListTileTheme.merge(
-            iconColor: _iconColor.value ?? expansionTileTheme.iconColor,
-            textColor: _headerColor.value,
-            child: InkWell(
-              customBorder: Border.all(width: 20.0, color: Colors.white),
-              borderRadius: widget.borderRadius,
-              onTap: _handleTap,
-              child: ListTile(
-                dense: true,
-                visualDensity: widget.isDefaultVerticalPadding
-                    ? null
-                    : const VisualDensity(horizontal: 0, vertical: -4),
-                contentPadding:
-                    widget.tilePadding ?? expansionTileTheme.tilePadding,
-                leading: widget.leading ?? _buildLeadingIcon(context),
-                title: widget.title,
-                subtitle: widget.subtitle,
-                trailing: widget.isHasTrailing == true
-                    ? widget.trailing ?? _buildTrailingIcon(context)
-                    : null,
-              ),
-            ),
-          ),
-          ClipRect(
-            child: Align(
-              alignment: widget.expandedAlignment ??
-                  expansionTileTheme.expandedAlignment ??
-                  Alignment.center,
-              heightFactor: _heightFactor.value,
-              child: child,
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-
-  @override
-  void didChangeDependencies() {
-    final ThemeData theme = Theme.of(context);
-    final ExpansionTileThemeData expansionTileTheme =
-        ExpansionTileTheme.of(context);
-    final ColorScheme colorScheme = theme.colorScheme;
-    _borderColorTween
-      ..begin =
-          widget.collapsedBorderColor ?? expansionTileTheme.backgroundColor
-      ..end = widget.expendedBorderColor ?? theme.dividerColor;
-    _headerColorTween
-      ..begin = widget.collapsedTextColor ??
-          expansionTileTheme.collapsedTextColor ??
-          theme.textTheme.titleMedium!.color
-      ..end = widget.textColor ??
-          expansionTileTheme.textColor ??
-          colorScheme.primary;
-    _iconColorTween
-      ..begin = widget.collapsedIconColor ??
-          expansionTileTheme.collapsedIconColor ??
-          theme.unselectedWidgetColor
-      ..end = widget.iconColor ??
-          expansionTileTheme.iconColor ??
-          colorScheme.primary;
-    _backgroundColorTween
-      ..begin = widget.collapsedBackgroundColor ??
-          expansionTileTheme.collapsedBackgroundColor
-      ..end = widget.backgroundColor ?? expansionTileTheme.backgroundColor;
-    super.didChangeDependencies();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final ExpansionTileThemeData expansionTileTheme =
-        ExpansionTileTheme.of(context);
-    final bool closed = !_isExpanded && _controller.isDismissed;
-    final bool shouldRemoveChildren = closed && !widget.maintainState;
-
-    final Widget result = Offstage(
-      offstage: closed,
-      child: TickerMode(
-        enabled: !closed,
-        child: Padding(
-          padding: widget.childrenPadding ??
-              expansionTileTheme.childrenPadding ??
-              EdgeInsets.zero,
-          child: Column(
-            crossAxisAlignment:
-                widget.expandedCrossAxisAlignment ?? CrossAxisAlignment.center,
-            children: widget.children,
-          ),
-        ),
-      ),
-    );
-
-    return AnimatedBuilder(
-      animation: _controller.view,
-      builder: _buildChildren,
-      child: shouldRemoveChildren ? null : result,
-    );
-  }
-}
